% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{bootstrapping}
\alias{bootstrapping}
\alias{basicConfig}
\alias{logReset}
\title{Bootstrapping the logging package.}
\usage{
basicConfig(level = 20)

logReset()
}
\arguments{
\item{level}{The logging level of the root logger. Defaults to INFO. Please do notice that
this has no effect on the handling level of the handler that basicConfig attaches to the
root logger.}
}
\description{
\code{basicConfig} and \code{logReset} provide a way to put the logging package
in a know initial state.
}
\details{
\code{basicConfig} creates the root logger, attaches a console handler(by
\var{basic.stdout} name) to it and sets the level of the handler to
\code{logging.level} INFO. The level of the logger can be passed as a
parameter to the function.

\code{logReset} reinitializes the whole logging system as if the package had just been
loaded. typically, you would want to call \code{basicConfig} immediately after a call
to \code{logReset}.
}
\examples{
basicConfig()
logdebug("not shown, basic is INFO")
logwarn("shown and timestamped")
logReset()
logwarn("not shown, as no handlers are present after a reset")

}
