% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{getLogger}
\alias{getLogger}
\title{Set defaults and get the named logger.}
\usage{
getLogger(name = "", ...)
}
\arguments{
\item{name}{The name of the logger}

\item{...}{Any properties you may want to set in the newly created
logger. These have no effect if the logger is already present.}
}
\value{
The logger retrieved or registered.
}
\description{
Make sure a logger with a specific name exists and return it as a
\var{Logger} S4 object.  if not yet present, the logger will be created and
given the values specified in the \dots arguments.
}
\examples{
getLogger()  # returns the root logger
getLogger('test.sub')  # constructs a new logger and returns it
getLogger('test.sub')  # returns it again

}
