% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedLevels.R
\docType{data}
\name{loglevels}
\alias{loglevels}
\title{The logging levels, names and values}
\format{An object of class \code{numeric} of length 11.}
\usage{
loglevels
}
\description{
This list associates names to values and vice versa.\cr
Names and values are the same as in the python standard logging module.
}
\keyword{datasets}
