% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{resetMsgComposer}
\alias{resetMsgComposer}
\title{Resets previously set message composer.}
\usage{
resetMsgComposer(container = "")
}
\arguments{
\item{container}{name of logger to reser message composer for (type: character)}
}
\description{
Resets previously set message composer.
}
