% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{setLevel}
\alias{setLevel}
\title{Set \var{logging.level} for the object.}
\usage{
setLevel(level, container = "")
}
\arguments{
\item{level}{The new level for this object. Can be numeric or character.}

\item{container}{a logger, its name or a handler. Default is root logger.}
}
\description{
Alter an existing logger or handler, setting its \var{logging.level} to a new
value. You can access loggers by name, while you must use \code{getHandler} to
get a handler.
}
\examples{
basicConfig()
setLevel("FINEST")
setLevel("DEBUG", getHandler("basic.stdout"))

}
